Move shardSpec tests to core
